# Do some more work upfront to speed up gc